requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pandas==2.1.3
pydantic==2.5.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.1

# Redis cache
redis==5.0.1

# PostgreSQL driver
psycopg2-binary==2.9.9

# Message queue
pika==1.3.2

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyotp==2.9.0
qrcode==7.4.2
cryptography==41.0.7

# WebSocket and Collaboration
python-socketio==5.10.0
aiohttp==3.9.1
y-py==0.6.2

# Machine Learning Framework
mlflow==2.9.2
torch==2.1.2
pytorch-lightning==2.1.3
autogluon==1.0.0
scikit-learn==1.3.2
shap==0.44.0
lime==0.2.0.1
numpy==1.24.3
scipy==1.11.4

# 3D Visualization
biopython==1.83
networkx==3.2.1
plotly==5.18.0
umap-learn==0.5.5

# Report Generation
jinja2==3.1.2
reportlab==4.0.7
python-docx==1.1.0
pylatex==1.4.2
pillow==10.1.0
matplotlib==3.8.2

# Statistical Analysis
lifelines==0.27.8
statsmodels==0.14.1
prophet==1.1.5
pymc==5.10.3
arviz==0.17.0
seaborn==0.13.0

# Distributed Processing
dask[complete]==2024.1.0
distributed==2024.1.0
bokeh==3.3.2

# Literature Mining and NLP
transformers==4.36.2
sentence-transformers==2.2.2
spacy==3.7.2
scispacy==0.5.3
neo4j==5.15.0
torch==2.1.2

# Plugin System
docker==7.0.0
bandit==1.7.5
safety==2.3.5
